Refactor / minimize code
