+++ wi-find.py
@@ -6,59 +6,79 @@
 # Copyright (C) 2019 Johnny Cache <johnycsh@gmail.com>
 #
 
-import sys
-import getopt
-import re
-import copy
-from colorama import Fore, Back, Style
+import sys, getopt, re, copy
 from scapy.all import *
-from rtap_ext_util import Listify_Radiotap_Headers
-from ext_beacon_util import rates_descriptor_t, modulation_descriptor_t, return_IE_by_tag, TargetCharacteristics
 
-from Rtap_Char import RadioTap_Profile_C, MeasureyM, Flatten_and_Average_MeasureM_list
+from rtap_ext_util import Listify_Radiotap_Headers
+from ext_beacon_util import return_IE_by_tag, TargetCharacteristics
+from Rtap_Char import  MeasureyM, Flatten_and_Average_MeasureM_list
+
+from colorama import Fore, Back, Style
 
 def Usage():
   print("Usage: %s -b <BSSID> -i <input>" % (sys.argv[0]))
   sys.exit(0)
 
-class Mathy_Stuff_Holder:
-
-  def Compare_dBm(self, a, b):
-  # For a good explanation of all these negative numbers:
-  # https://community.cisco.com/t5/small-business-support-documents/why-is-almost-everything-negative-in-wireless/ta-p/3159743
-  # 1dBm = 1.258925 mW
-
-    deltaDbm = 1.0 * abs(a - b)
-    #print("    deltaDbm(%d,%d) = %d" % (a,b, deltaDbm))
-    return deltaDbm
-
-  ### Comparing dBm: How do i always forget this formula?
-  ### Quick note: The 'old' version (ap-finder.c) simply told you if curr_avg was (Better, equal, Worse) than prev_avg
-  ###             This actually does not require us to work/convert into/from a logarithmic scale.
-  ###             -50 is better than -51. The Distance is irrelvant (sortof.)
-  def RateSelf(self):
-    #print("    RateSelf(%3d,%3d)" % (self.prev_avg, self.curr_avg))
-    print("    Delta_dBm (%3d, %3d) = %2d" % ( self.State.curr_avg,  self.State.prev_avg, self.Compare_dBm(self.State.curr_avg, self.State.prev_avg)))
-    print("    Delta_dBm (%3d, %3d) = %2d " % ( self.State.curr_avg, self.Config.Ref_dBm, self.Compare_dBm(self.State.curr_avg, self.Config.Ref_dBm)))
-  ### TODO: Really should make some sort of BeaconMeasurement() class that handles dBm/ SNR / .. conversion and comparison
-
-
 
 class ConfigC:  #Set-once configuration paramters (values do not change during main loop)
   BSSID=None
-  SSID=None
   input_src=None    # 'en0', file.pcap, ...
   sniff_mode=None #  Valid options: 'offline' (pcap file) or 'iface' (self-descr)
-  Ref_dBm=-40   #Pick an arbitrary (but consistent) 'standard candle'
-
   pkts_per_avg=3
+  
+  def Parse_Args(self):
+      print("#### Parse_Args: Start")
+      opts = getopt.getopt(sys.argv[1:],"b:i:h")
+
+      for opt,optarg in opts[0]:
+        if opt == "-b":
+          self.BSSID = optarg
+        elif opt == "-i":
+          self.input_src = optarg
+        elif opt == "-h":
+          Usage()
+
+      if not self.BSSID:
+        print("\nError: BSSID not defined\n")
+        Usage()
+      if re.match('^([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$', self.BSSID):
+        self.BSSID = self.BSSID.lower()
+      else:
+        print("\nError: Wrong format for BSSID\n")
+        Usage()
+
+      if not (self.input_src):
+        print("\nError: Input not specified")
+        Usage()
+
+      # Attempt to open input as both file and iface
+      try:
+        rdpcap(filename=self.input_src, count=1)
+      except:
+        pass
+      else:
+        print("Parse_args: %s opened as file success" % (self.input_src))
+        self.sniff_mode="offline"
 
 
-### JC: TODO: 
-### In the global State variable StateC, add:
-### StateC.AntennaMeasureList = []
-##### StateC.AntennaMeasureList[BitB] = list(measurements for Rtap field BitB)
+      if (self.sniff_mode != "offline"):
+        try:
+          print("asdf")
+          sniff(iface=self.input_src, monitor=1, store=0, count=1)
 
+        except:
+          print("Error. %s not valid as input file or interface. Exiting." % (self.input_src))
+          sys.exit(0)
+        else:
+          print("Opened input src as iface successfully.")
+          self.sniff_mode="iface"
+
+      if self.sniff_mode == "offline":
+        print("    ---- Offline mode enabled")
+      elif self.sniff_mode == "iface":
+        print("     ------Online mode enabled")
+      else:
+        Usage()
 
 class StateC:  #All dynamic state associated with instance
   cnt=0
@@ -71,21 +91,6 @@ class StateC:  #All dynamic state associated with instance
   ###Extended Radiotap records will be stored at AnteannaId+1. 
   ### //This accounts for the fact that 'Extended' Antenna rtap headers typically start
   ### with AntennaID 0.
-  ### So:
-  ### AntennaMeasureList[0] = TopLevelRadiotapHeader
-  ### AntennaMeasureList[1] = AntennaId.0
-  ### AntennaMeasureList[2] = AntennaID.1, ..
-  ### AntennaMeasureList[3] = AntennaId.2
-  ### AntennaMeasureList[4] = AntennaId.3
-  ### AntennaMeasureList[X] = 1-Dimensional list, offset=AntennaId
-  ### AntennaMeasureList[X].MtoL[RtapBitId] = list()
-  ### AntennaMeasureList[0] = MeasureyM
-  ### AntennaMeasurey_Ma
-  ### Measurey_Map
-  ### MeasureyM.update(R)
-  ### ## M.cnt+=1
-  ### ## M.AntSignal_List = []
-
   
   curr_avg_sig = 0
   prev_avg_sig = 0
@@ -95,115 +100,27 @@ class StateC:  #All dynamic state associated with instance
   prev_avg_noise = 0
   curr_noisedBms = []
 
-
   pkt_measurements_curr = []
   avg_pkt_measurements_curr = None
   pkt_measurements_prev = None
   avg_pkt_measurements_prev = None
+
   
-
-  #curr_measurement_avg = MeasureyM()
-
-  ### Set of _Minimum_ fields present
-
-  ### Clever: create a dictionary that maps the fields bit-positin in RadioTap Present -> list of stored values
-  present_bits_analyzed = [6,7] #//Sig_dBm,Noise_dBm
-  readings = {}
-
-  curr_c = 0     #TODO Replace with call to len
-
-  num_times_before_reprinting_network_header=3
-
 
   def init(self):
     print("#### StateC::ctor::Start")
     input("Continue:")
-
-
-
-
 
 class MainAppC:
 
   State = StateC()
   Config = ConfigC()
   Target = TargetCharacteristics()
-  def Parse_Args(self):
-      print("#### Parse_Args: Start")
-      opts = getopt.getopt(sys.argv[1:],"b:i:h")
-
-      for opt,optarg in opts[0]:
-        if opt == "-b":
-          self.Config.BSSID = optarg
-        elif opt == "-i":
-          self.Config.input_src = optarg
-        elif opt == "-h":
-          Usage()
-
-      if not self.Config.BSSID:
-        print("\nError: BSSID not defined\n")
-        Usage()
-      if re.match('^([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$', self.Config.BSSID):
-        self.Config.BSSID = self.Config.BSSID.lower()
-      else:
-        print("\nError: Wrong format for BSSID\n")
-        Usage()
-
-      if not (self.Config.input_src):
-        print("\nError: Input not specified")
-        Usage()
-
-      # Attempt to open input as both file and iface
-      try:
-        rdpcap(filename=self.Config.input_src, count=1)
-      except:
-        pass
-      else:
-        print("Parse_args: %s opened as file success" % (self.Config.input_src))
-        self.Config.sniff_mode="offline"
-
-
-      if (self.Config.sniff_mode != "offline"):
-        try:
-          print("asdf")
-          sniff(iface=self.Config.input_src, monitor=1, store=0, count=1)
-
-        except:
-          print("Error. %s not valid as input file or interface. Exiting." % (self.Config.input_src))
-          sys.exit(0)
-        else:
-          print("Opened input src as iface successfully.")
-          self.Config.sniff_mode="iface"
-
-      if self.Config.sniff_mode == "offline":
-        print("    ---- Offline mode enabled")
-      elif self.Config.sniff_mode == "iface":
-        print("     ------Online mode enabled")
-      else:
-        Usage()
-
-
-
-
-  def Careful_Process_Radiotap_(self, pkt):
-    reqd_rtap_pres = ['Rate', 'Channel', 'dBm_AntSignal'] #List of minimum set of viable radiotap fields to be useful
-    self.State.cnt+=1
-    print("--%2d): #### AntTuner::ProcessRadiotap" % (self.State.cnt))
-    if ( not pkt.haslayer(RadioTap)):
-      print("    No RadioTap header(s) detected")
-      return
-
-    #rtap=pkt[RadioTap]
-
-    print("    Present: 0x%08x: " % (rtap.present))
-    for k in self.reqd_rtap_pres:
-      print("Good: %s Marked present" %(k))
-
-    print("    Rate: %d Channel:%d dBm_AntSignal: %d  Lock_Quality: %d" % (rtap.Rate, rtap.Channel,  rtap.dBm_AntSignal, rtap.Lock_Quality))
+  
 
   def callback_main(self, pkt):
     self.State.cnt+=1
-    header_list=[]
+    #header_list=[]
     argslist=[]
     colorlist=[]
     print("--%2d): #### sniff::callback_main" % (self.State.cnt))
@@ -275,15 +192,6 @@ class MainAppC:
     ##### ---- Begin dynamic format string creation ----- #######
     ##Walk the present bitmask, generating a header list and a value list in parallel
     
-    # 
-    
-#     print(fmt_str % map(str, argslist))
-
-
-
-    #self.State.prev_avg_noise =self.State.curr_avg_noise
-    #self.State.curr_avg_noise = sum(self.State.curr_noisedBms) / pkts_per_avg
-
 
     if (self.State.curr_avg_sig == self.State.prev_avg_sig):
       curr_color = Fore.WHITE
@@ -303,12 +211,6 @@ class MainAppC:
     #self.RateSelf()
 
 
-
-
-
-    
-
-
 def ParseTargetBeacon(pkt):
 
   print("----- Analyzing beacon into target characteristics")
@@ -316,8 +218,7 @@ def ParseTargetBeacon(pkt):
   T.init(pkt)
   print("-----  Target Summary above ----")
   input("")
-  #ssid=str(pkt.getlayer(Dot11).info) #XXX This conveniently contains SSID (IELement 0. But this isnt a great approach)
-
+  
 def main():
   ## Misc platform setup: On Macos we need to explicitly enable libpcap for BPF to work correctly
 
@@ -325,8 +226,7 @@ def main():
   #YYY: TODO wrap sniff() calls in conf.use_pcap cases
 
   A = MainAppC()
-  A.Parse_Args()
-  C = RadioTap_Profile_C()
+  A.Config.Parse_Args()
 
   bpfilter="type mgt and subtype beacon and wlan host %s " % ( A.Config.BSSID)
   ### Before we can get to the main loop, we need to catch atleast 1 beacon (so we know how many measurements are present etc)
@@ -349,9 +249,6 @@ def main():
     sniff(prn=A.callback_main, offline=A.Config.input_src, filter=bpfilter, monitor=1, store=0, count=0)
   else:
      sniff(prn=A.callback_main, iface=A.Config.input_src, filter=bpfilter, monitor=1, store=0, count=0)
-
-
-
 
 if __name__=='__main__':
   main()
